chore: Write long inputs to file
Prior to this, we ignored the tests that had very long inputs. So far, 1 test per variant.

These long inputs could not be in a java file.

This change writes these long inputs to a file and reads from the file in the test instead of inlining in the test.